refactor : cicd update
