--- .github/workflows/gradle.yml
@@ -1,108 +1,0 @@
-# This workflow uses actions that are not certified by GitHub.
-# They are provided by a third-party and are governed by
-# separate terms of service, privacy policy, and support
-# documentation.
-# This workflow will build a Java project with Gradle and cache/restore any dependencies to improve the workflow execution time
-# For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-java-with-gradle
-
-name: Java CI with Gradle
-
-on:
-  push:
-    branches: [ "dev" ]
-  pull_request:
-    branches: [ "dev" ]
-
-jobs:
-  build:
-
-    runs-on: ubuntu-latest
-    permissions:
-      contents: read
-
-    steps:
-    - uses: actions/checkout@v4
-    - name: Set up JDK 21
-      uses: actions/setup-java@v4
-      with:
-        java-version: '21'
-        distribution: 'temurin'
-
-    # Configure Gradle for optimal use in GitHub Actions, including caching of downloaded dependencies.
-    # See: https://github.com/gradle/actions/blob/main/setup-gradle/README.md
-    - name: Setup Gradle
-      uses: gradle/actions/setup-gradle@af1da67850ed9a4cedd57bfd976089dd991e2582 # v4.0.0
-      with:
-        cache-disabled: true
-        build-scan-publish: false
-
-    - name: Grant execute permission for gradlew
-      working-directory: ./back
-      run: chmod +x gradlew
-
-    - name: Change to back directory
-      run: cd back
-      
-    - name: Build with Gradle Wrapper
-      working-directory: ./back
-      env:
-        JWT_SECRET: ${{ secrets.JWT_SECRET }}
-        S3STORAGE_ACCESS_KEY: ${{ secrets.S3STORAGE_ACCESS_KEY }}
-        S3STORAGE_SECRET_KEY: ${{ secrets.S3STORAGE_SECRET_KEY }}
-        TOSS_SECRET_KEY: ${{ secrets.TOSS_SECRET_KEY }}
-      run: ./gradlew build
-
-    - name: Check build results
-      working-directory: ./back
-      run: ls -la build/libs/
-
-    - name: Upload JAR file
-      uses: appleboy/scp-action@master
-      with:
-        host: ${{ secrets.SSH_IP }}
-        username: ${{ secrets.SSH_ID }}
-        key: ${{ secrets.SSH_KEY }}
-        # password: ${{ secrets.SSH_PASSWORD }}
-        port: ${{ secrets.SSH_PORT }}
-        source: "back/build/libs/*.jar"
-        target: "~/app"
-        rm: false
-        
-    - name: Execute shell script
-      uses: appleboy/ssh-action@master
-      with:
-        host: ${{ secrets.SSH_IP }}
-        username: ${{ secrets.SSH_ID }}
-        key: ${{ secrets.SSH_KEY }}
-        # password: ${{ secrets.SSH_PASSWORD }}
-        port: ${{ secrets.SSH_PORT }}
-        script_stop: true
-        script: "~/startup.sh"
-
-## dependency-submisttion job 제거
-  # dependency-submission:
-  #   runs-on: ubuntu-latest
-  #   permissions:
-  #     contents: write
-  #   steps:
-  #   - uses: actions/checkout@v4
-    
-  #   - name: Set up JDK 21
-  #     uses: actions/setup-java@v4
-  #     with:
-  #       java-version: '21'
-  #       distribution: 'temurin'
-    
-  #   # 디렉토리 이동 및 권한 부여
-  #   - name: Setup Gradle Wrapper
-  #     run: |
-  #       cd back
-  #       chmod +x gradlew
-    
-  #   # gradle-version 사용하거나 build-root-directory 사용
-  #   - name: Generate and submit dependency graph
-  #     uses: gradle/actions/dependency-submission@af1da67850ed9a4cedd57bfd976089dd991e2582
-  #     with:
-  #       build-root-directory: back
-  #       # 또는 Gradle 버전 직접 지정 (Wrapper 대신)
-  #       # gradle-version: '8.5'
