+++ back/build.gradle
@@ -21,7 +21,7 @@ configurations {
 }
 
 jacoco {
-    toolVersion = "0.8.8"
+    toolVersion = "0.8.11"
 }
 
 jacocoTestReport {
