+++ back/src/main/java/com/bubble/giju/domain/order/controller/OrderController.java
@@ -1,6 +1,5 @@
 package com.bubble.giju.domain.order.controller;
 
-import com.bubble.giju.domain.cart.dto.response.CartListResponseDto;
 import com.bubble.giju.domain.order.dto.request.DirectOrderRequestDto;
 import com.bubble.giju.domain.order.dto.request.OrderRequestDto;
 import com.bubble.giju.domain.order.dto.request.RefundRequestDto;
@@ -31,7 +30,6 @@ import java.util.List;
 public class OrderController {
 
     private final OrderService orderService;
-
 
     @Operation(summary = "장바구니 담고 주문 생성", description = "order, orderDetail 생성, toss연결을 위한 orderId, amount, orderName, customerName, customerEmail, successUrl, failUrl 반환")
     @PostMapping
