+++ back/src/main/java/com/bubble/giju/domain/order/dto/request/OrderRequestDto.java
@@ -1,10 +1,7 @@
 package com.bubble.giju.domain.order.dto.request;
 
-import com.bubble.giju.domain.order.entity.OrderStatus;
 import lombok.AllArgsConstructor;
-import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 import java.util.List;
