+++ back/src/main/java/com/bubble/giju/domain/order/dto/response/OrderHistoryResonseDto.java
@@ -4,7 +4,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
 import java.util.List;
