+++ back/src/main/java/com/bubble/giju/domain/order/entity/OrderDetail.java
@@ -1,7 +1,5 @@
 package com.bubble.giju.domain.order.entity;
 
-
-import com.bubble.giju.domain.user.entity.User;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Builder;
